+++ Sources/ProjectSpec/TargetSource.swift
@@ -23,8 +23,8 @@ public struct TargetSource: Equatable {
     public var createIntermediateGroups: Bool?
     public var attributes: [String]
     public var resourceTags: [String]
-    public var inferPlatformFiltersByPath: Bool?
-    public var platformFilters: [SupportedDestination]?
+    public var inferDestinationFiltersByPath: Bool?
+    public var destinationFilters: [SupportedDestination]?
 
     public enum HeaderVisibility: String {
         case `public`
@@ -54,8 +54,8 @@ public struct TargetSource: Equatable {
         createIntermediateGroups: Bool? = nil,
         attributes: [String] = [],
         resourceTags: [String] = [],
-        inferPlatformFiltersByPath: Bool? = nil,
-        platformFilters: [SupportedDestination]? = nil
+        inferDestinationFiltersByPath: Bool? = nil,
+        destinationFilters: [SupportedDestination]? = nil
     ) {
         self.path = (path as NSString).standardizingPath
         self.name = name
@@ -70,8 +70,8 @@ public struct TargetSource: Equatable {
         self.createIntermediateGroups = createIntermediateGroups
         self.attributes = attributes
         self.resourceTags = resourceTags
-        self.inferPlatformFiltersByPath = inferPlatformFiltersByPath
-        self.platformFilters = platformFilters
+        self.inferDestinationFiltersByPath = inferDestinationFiltersByPath
+        self.destinationFilters = destinationFilters
     }
 }
 
@@ -119,10 +119,10 @@ extension TargetSource: JSONObjectConvertible {
         attributes = jsonDictionary.json(atKeyPath: "attributes") ?? []
         resourceTags = jsonDictionary.json(atKeyPath: "resourceTags") ?? []
         
-        inferPlatformFiltersByPath = jsonDictionary.json(atKeyPath: "inferPlatformFiltersByPath")
+        inferDestinationFiltersByPath = jsonDictionary.json(atKeyPath: "inferDestinationFiltersByPath")
         
-        if let platformFilters: [SupportedDestination] = jsonDictionary.json(atKeyPath: "platformFilters") {
-            self.platformFilters = platformFilters
+        if let destinationFilters: [SupportedDestination] = jsonDictionary.json(atKeyPath: "destinationFilters") {
+            self.destinationFilters = destinationFilters
         }
     }
 }
@@ -141,8 +141,8 @@ extension TargetSource: JSONEncodable {
             "createIntermediateGroups": createIntermediateGroups,
             "resourceTags": resourceTags,
             "path": path,
-            "inferPlatformFiltersByPath": inferPlatformFiltersByPath,
-            "platformFilters": platformFilters?.map { $0.rawValue },
+            "inferDestinationFiltersByPath": inferDestinationFiltersByPath,
+            "destinationFilters": destinationFilters?.map { $0.rawValue },
         ]
 
         if optional != TargetSource.optionalDefault {