+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 - Added `supportedDestinations` for target
 - Added a new platform value `auto` that we can use only with `supportedDestinations`
 - Added the possiblity to avoid the definition of plaform, only when we use `supportedDestinations`, that fallbacks to 'auto'
-- Added `platformFilters` for sources and dependecies
-- Added `inferPlatformFiltersByPath`, a convenience filter for sources
+- Added `destinationFilters` for sources and dependecies
+- Added `inferDestinationFiltersByPath`, a convenience filter for sources
 - Added visionOS support
 
 ### Added